update Bitcointalk tests
