+++ crypto-scraper/tests/Unit/Bitcointalk/GetBoardPagesTest.php
@@ -1,8 +1,8 @@
 <?php
 
-namespace Tests\Unit;
+namespace Tests\Unit\Bitcointalk;
 
-use App\Console\Commands\BitcointalkLoadBoards;
+use App\Console\Commands\Bitcointalk\LoadBoards;
 use Tests\TestCase;
 
 class GetBoardPagesTest extends TestCase
@@ -18,7 +18,7 @@ class GetBoardPagesTest extends TestCase
             'https://bitcointalk.org/index.php?board=22.33',
             'https://bitcointalk.org/index.php?board=22222.23453453234'
         ];
-        $output = BitcointalkLoadBoards::getBoardPages($input);
+        $output = LoadBoards::getBoardPages($input);
         $this->assertEquals($input, $output);
     }
     
@@ -33,14 +33,14 @@ class GetBoardPagesTest extends TestCase
             'https://bitcointalk.org/index.php?board=22.234543',
             'https://bitcointalk.org/index.php?board=22.1234'
         ];
-        $output = BitcointalkLoadBoards::getBoardPages($input);
+        $output = LoadBoards::getBoardPages($input);
         $this->assertEquals($expected, $output);
     }
     
     public function testEmptyBoardPages() {
         $input = [];
         $expected = [];
-        $output = BitcointalkLoadBoards::getBoardPages($input);
+        $output = LoadBoards::getBoardPages($input);
         $this->assertEquals($expected, $output);
     }
     
@@ -52,7 +52,7 @@ class GetBoardPagesTest extends TestCase
         $expected = [
             'https://bitcointalk.org/index.php?board=2.2345'
         ];
-        $output = BitcointalkLoadBoards::getBoardPages($input);
+        $output = LoadBoards::getBoardPages($input);
         $this->assertEquals($expected, $output);
     }
 }
