+++ crypto-scraper/tests/Unit/Bitcointalk/MainBoardValidTest.php
@@ -1,8 +1,8 @@
 <?php
 
-namespace Tests\Unit;
+namespace Tests\Unit\Bitcointalk;
 
-use App\Console\Commands\BitcointalkLoadBoards;
+use App\Console\Commands\Bitcointalk\LoadBoards;
 use Tests\TestCase;
 
 class MainBoardValidTest extends TestCase
@@ -13,17 +13,17 @@ class MainBoardValidTest extends TestCase
      * @return void
      */
     public function testValid() {
-        $output = BitcointalkLoadBoards::mainBoardValid('https://bitcointalk.org/index.php?board=83.0');
+        $output = LoadBoards::mainBoardValid('https://bitcointalk.org/index.php?board=83.0');
         $this->assertEquals(true, $output);
     }
     
     public function testBoardPage() {
-        $output = BitcointalkLoadBoards::mainBoardValid('https://bitcointalk.org/index.php?board=83.1');
+        $output = LoadBoards::mainBoardValid('https://bitcointalk.org/index.php?board=83.1');
         $this->assertEquals(false, $output);
     }
     
     public function testBrokenMainBoard() {
-        $output = BitcointalkLoadBoards::mainBoardValid('https://bitcointalk.org/index.php?board=83.0asdfsdf');
+        $output = LoadBoards::mainBoardValid('https://bitcointalk.org/index.php?board=83.0asdfsdf');
         $this->assertEquals(false, $output);
     }
 }
